enum class FriendStatus

+++ Source/_Database/DbManagerSave.cpp
@@ -228,7 +228,7 @@ int MAX_SAVEPARAM = 88;
 #else	// __JEFF_9_20
 	bOK[++j] = qry->BindParameter( ++i, SQL_PARAM_INPUT, SQL_C_LONG, SQL_INTEGER,   0, 0, &pMover->m_dwFlyTime, 0, 0 );
 #endif	// __JEFF_9_20
-	DWORD dwState	= pMover->m_RTMessenger.GetState();
+	long dwState	= static_cast<long>(pMover->m_RTMessenger.GetState());
 	bOK[++j] = qry->BindParameter( ++i, SQL_PARAM_INPUT, SQL_C_LONG, SQL_INTEGER,   0, 0, &dwState, 0, 0 );
 	bOK[++j] = qry->BindParameter( ++i, SQL_PARAM_INPUT, SQL_C_LONG, SQL_INTEGER,   0, 0, &nTotalPlayTime, 0, 0 );
 	bOK[++j] = qry->BindParameter( ++i, SQL_PARAM_INPUT, SQL_C_LONG, SQL_INTEGER,   0, 0, &pMover->m_nPKValue, 0, 0 );
